+++ packages/expo-notifications/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### 💡 Others
 
+## 0.30.1 — 2025-04-08
+
+_This version does not introduce any user-facing changes._
+
 ## 0.30.0 — 2025-04-04
 
 ### 🐛 Bug fixes